Added virtual joystick support (install PPjoy)

+++ src/main/kotlin/nl/sajansen/hardwarecockpitclient/hardware/serial/SerialOperationMode.kt
@@ -1,11 +1,11 @@
 package nl.sajansen.hardwarecockpitclient.hardware.serial
 
-enum class SerialOperationMode(val id: Int) {
+enum class SerialOperationMode(val id: Byte) {
     OPERATION_MODE_UNCONNECTED(0),
     OPERATION_MODE_SIMULATOR(1),
     OPERATION_MODE_CALIBRATION(2);
 
     companion object {
-        fun fromInt(value: Int) = values().first { it.id == value }
+        fun fromInt(value: Int) = values().first { it.id == value.toByte() }
     }
 }
